Move variable subscriptions out of QuickPrayers.kts
