+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/prayer/QuickPrayers.kts
@@ -118,30 +118,6 @@ fun Player.togglePrayer(prayerIndex: Int, listKey: String, quick: Boolean) {
     }
 }
 
-val variables: VariableStore by inject()
-
-on<VariableSet>({ key == ACTIVE_PRAYERS || key == ACTIVE_CURSES }) { player: Player ->
-    val from = from as Int
-    val to = to as Int
-    val variable = variables.get(key) as BitwiseVariable<String>
-    for (id in variable.values) {
-        val value = variable.getValue(id) ?: continue
-        if (from.has(value) && !to.has(value)) {
-            player.events.emit(PrayerDeactivate(id, key == ACTIVE_CURSES))
-        } else if (!from.has(value) && to.has(value)) {
-            player.events.emit(PrayerActivate(id, key == ACTIVE_CURSES))
-        }
-    }
-}
-
-on<VariableAdded>({ key == ACTIVE_PRAYERS || key == ACTIVE_CURSES }) { player: Player ->
-    player.events.emit(PrayerActivate(value as String, key == ACTIVE_CURSES))
-}
-
-on<VariableRemoved>({ key == ACTIVE_PRAYERS || key == ACTIVE_CURSES }) { player: Player ->
-    player.events.emit(PrayerDeactivate(value as String, key == ACTIVE_CURSES))
-}
-
 /**
  * Quick prayers
  * Until the new quick prayer selection is confirmed old
